ToDo page Sprint 2 done
Uploading on behalf of Tan Jun Hong.
To-Do items filtered to Users. i.e. Only To-Do items created by Owner is shown.
Addition and deletion of To-Do items auto archived instead of the current manual process.
All To-Do items to have a timestamp added upon creation automatically by the application.